move `LinkSelfContainedComponents` to `rustc_target`
